dhofar_flavor_chat project/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

langchain
langchain-community
langchain-core
langchain-ollama

faiss-cpu
pydantic